requirements.txt
fastapi
uvicorn[standard]
pydantic
ollama
slowapi
python-dotenv